third_party/blink/web_tests/css3/supports-expected.txt

Test the @supports rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById('t0')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t1')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t2')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t3')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t4')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t5')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t6')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t7')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t8')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t9')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t10')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t11')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t12')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t13')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t14')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t15')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t16')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t17')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t18')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t19')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t20')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t21')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t22')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t23')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t24')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t25')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t26')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t27')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t28')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t29')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t30')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t31')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t32')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t33')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t34')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t35')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t36')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t37')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t38')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t39')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t40')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t41')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t42')).content is "\"APPLIED\""
PASS getComputedStyle(document.getElementById('t43')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t44')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t45')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t46')).content is "\"UNTOUCHED\""
PASS getComputedStyle(document.getElementById('t47')).content is "\"APPLIED\""
PASS successfullyParsed is true

TEST COMPLETE
